Adding homepage background image
I found this stock image on Unsplash which features two hijabis surrounded by lush green scenery. This is simply a temporary stand-in for a better one in the future. Preferably, I'd like an MSA group photo OR a background that matches our white, green, and black MSA theme. I hope I can discuss this with the team soon.