Update index.html
Assalamu'Alaikum again!
List of things that have been changed:
Navigation button text for (only for Home and Mission)
home page background image
Home page text
home page button text which links to Mission statement
mission statement page text
mission statement button which links to services
Changed the theme of the website background from Yellow to green (please change if not to the tech team's taste. I know I only had the Home page and Mission statement to do, but yellow really does not fit with our MSA logo).

Now, this is definitely a work in progress. The mission statement is from lines I obtained from the QCMSA Facebook and Twitter pages but the page felt kind of empty with just two lines, so I added another sentence which I feel transitions to the "services" page. 
Also, as I've mentioned in the img pull request, we need to change the homepage background image. This is a generic image online, but it would be better if we found an image that reps our MSA, fits the color theme of our MSA logo, or even better, both! Let me know what I can do.